TalentScout AI Hiring Assistant/requirements.txt
streamlit
openai
python-dotenv 